Improve readers query

+++ src/GoogleBooks.Application/Common/Models/EntitiesByCriteria.cs
@@ -2,7 +2,7 @@
 
 public class EntitiesByCriteria<TEntities>
 {
-    public int TotalItems { get; set; }
+    public long TotalItems { get; set; }
 
     public TEntities[] Items { get; set; } = default!;
 }
